restructure test function
