+++ pkg/v1/tkg/test/tkgctl/shared/e2e_common.go
@@ -10,11 +10,11 @@ import (
 	"math/rand"
 	"os"
 	"path/filepath"
-	"strings"
 	"time"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/cluster-api/util"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
@@ -22,15 +22,7 @@ import (
 	"github.com/vmware-tanzu/tanzu-framework/pkg/v1/tkg/test/framework"
 	"github.com/vmware-tanzu/tanzu-framework/pkg/v1/tkg/tkgctl"
 
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	secretutil "sigs.k8s.io/cluster-api/util/secret"
-
-	kappctrl "github.com/vmware-tanzu/carvel-kapp-controller/pkg/apis/kappctrl/v1alpha1"
-	kapppkgiv1alpha1 "github.com/vmware-tanzu/carvel-kapp-controller/pkg/apis/packaging/v1alpha1"
-	addonutil "github.com/vmware-tanzu/tanzu-framework/addons/pkg/util"
-	runtanzuv1alpha3 "github.com/vmware-tanzu/tanzu-framework/apis/run/v1alpha3"
-	clusterapiv1beta1 "sigs.k8s.io/cluster-api/api/v1beta1"
+	pkgiv1alpha1 "github.com/vmware-tanzu/carvel-kapp-controller/pkg/apis/packaging/v1alpha1"
 )
 
 type E2ECommonSpecInput struct {
@@ -164,7 +156,10 @@ func E2ECommonSpec(context context.Context, inputGetter func() E2ECommonSpecInpu
 		framework.WaitForNodes(framework.NewClusterProxy(clusterName, tempFilePath, ""), 2)
 
 		By(fmt.Sprintf("Verify addon packages on workload cluster %q matches clusterBootstrap info on management cluster", clusterName))
-		err = checkUtkgAddons(context, client, scheme, "antrea")
+		scheme := runtime.NewScheme()
+		err = pkgiv1alpha1.AddToScheme(scheme)
+		Expect(err).NotTo(HaveOccurred())
+		err = checkPackageInstalls(context, client, scheme, "antrea")
 		Expect(err).To(BeNil())
 
 		By(fmt.Sprintf("Deleting workload cluster %q", clusterName))
@@ -177,106 +172,4 @@ func E2ECommonSpec(context context.Context, inputGetter func() E2ECommonSpecInpu
 
 		By("Test successful !")
 	})
-}
-
-// getClusterBootstrap gets ClusterBootstrap resource with the provided object key
-func getClusterBootstrap(ctx context.Context, k8sClient client.Client, namespace, clusterName string) *runtanzuv1alpha3.ClusterBootstrap {
-	clusterBootstrap := &runtanzuv1alpha3.ClusterBootstrap{}
-	objKey := client.ObjectKey{Namespace: namespace, Name: clusterName}
-
-	Eventually(func() bool {
-		err := k8sClient.Get(ctx, objKey, clusterBootstrap)
-		return err == nil
-	}, waitTimeout, pollingInterval).Should(BeTrue())
-
-	Expect(clusterBootstrap).ShouldNot(BeNil())
-	return clusterBootstrap
-}
-
-// getPackageInstall get PackageInstall resource with the provided object key
-func getPackageInstall(ctx context.Context, k8sClient client.Client, namespace, pkgiName string) *kapppkgiv1alpha1.PackageInstall {
-	pkgInstall := &kapppkgiv1alpha1.PackageInstall{}
-	objKey := client.ObjectKey{Namespace: namespace, Name: pkgiName}
-
-	Eventually(func() bool {
-		err := k8sClient.Get(ctx, objKey, pkgInstall)
-		return err == nil
-	}, waitTimeout, pollingInterval).Should(BeTrue())
-
-	Expect(pkgInstall).ShouldNot(BeNil())
-
-	return pkgInstall
-}
-
-func getPackageDetailsFromCBS(CBSRefName string) (string, string, string, error) {
-	pkgShortName := strings.Split(CBSRefName, ".")[0]
-
-	pkgName := strings.Join(strings.Split(CBSRefName, ".")[0:4], ".")
-
-	pkgVersion := strings.Join(strings.Split(CBSRefName, ".")[4:], ".")
-
-	return pkgShortName, pkgName, pkgVersion, nil
-}
-
-func checkUtkgAddons(ctx context.Context, cl client.Client, scheme *runtime.Scheme, testPkgName string) error {
-	var (
-		err          error
-		pkgShortName string
-		pkgName      string
-		pkgVersion   string
-	)
-	mngCluster := &clusterapiv1beta1.Cluster{}
-
-	// create remoteClient for workload cluster
-	remoteClient, err := addonutil.GetClusterClient(ctx, cl, scheme, util.ObjectKey(mngCluster))
-	if err != nil {
-	    return err
-	}
-
-	key := client.ObjectKey{Namespace: constants.TkgNamespace, Name: secretutil.Name(mngCluster.Name, secretutil.Kubeconfig)}
-	clusterKubeConfigSecret := &corev1.Secret{}
-	err = remoteClient.Get(ctx, key, clusterKubeConfigSecret)
-	if err != nil {
-		return err
-	}
-
-	//clusterBootstrap := getClusterBootstrap(client.ObjectKeyFromObject(mngCluster))
-
-	// Get ClusterBootstrap and return error if not found
-	clusterBootstrap := getClusterBootstrap(ctx, cl, constants.TkgNamespace, mngCluster.Name)
-
-	//wlcCluster := &clusterapiv1beta1.Cluster{}
-
-	// packageInstall name for for both management and workload clusters should follow the <cluster name>-<addon short name>
-	// packageInstall name and version should match info in clusterBootstrap for all packages, format is <package name>.<package version>
-	switch {
-	case testPkgName == "CNI":
-		pkgShortName, pkgName, pkgVersion, err = getPackageDetailsFromCBS(clusterBootstrap.Spec.CNI.RefName)
-	case testPkgName == "CSI":
-		pkgShortName, pkgName, pkgVersion, err = getPackageDetailsFromCBS(clusterBootstrap.Spec.CSI.RefName)
-	case testPkgName == "CPI":
-		pkgShortName, pkgName, pkgVersion, err = getPackageDetailsFromCBS(clusterBootstrap.Spec.CPI.RefName)
-	case testPkgName == "Kapp":
-		pkgShortName, pkgName, pkgVersion, err = getPackageDetailsFromCBS(clusterBootstrap.Spec.Kapp.RefName)
-	case testPkgName == "metrics-server":
-		pkgShortName, pkgName, pkgVersion, err = getPackageDetailsFromCBS(clusterBootstrap.Spec.AdditionalPackages[0].RefName)
-	case testPkgName == "secretgen-controller":
-		pkgShortName, pkgName, pkgVersion, err = getPackageDetailsFromCBS(clusterBootstrap.Spec.AdditionalPackages[1].RefName)
-	case testPkgName == "pinniped":
-		pkgShortName, pkgName, pkgVersion, err = getPackageDetailsFromCBS(clusterBootstrap.Spec.AdditionalPackages[2].RefName)
-	}
-
-	pkgiName := addonutil.GeneratePackageInstallName(mngCluster.Name, pkgShortName)
-	pkgi := getPackageInstall(ctx, remoteClient, constants.TkgNamespace, pkgiName)
-	// check package install reconcile status is succeed
-	Expect(pkgi.Status.GenericStatus.Conditions[1].Type).Should(Equal(kappctrl.ReconcileSucceeded))
-	Expect(pkgi.Status.GenericStatus.Conditions[1].Status).Should(Equal(corev1.ConditionTrue))
-
-	// Verify package name match between clusterBootstrap and packageInstall
-	Expect(pkgName).Should(Equal(pkgi.Spec.PackageRef.RefName))
-
-	// Verify package version match between clusterBootstrap and packageInstall
-	Expect(pkgVersion).Should(Equal(pkgi.Spec.PackageRef.VersionSelection.Constraints))
-
-	return nil
 }
